backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
google-generativeai==0.3.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.1
aiofiles==23.2.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
requests==2.32.3
jinja2==3.1.3
openpyxl==3.1.5
